Feature/step 1 poc (#1)
* [add] simple mongodb
* [add] basic setup and requirements
* [add] simple test
* [add] simple app
* [add] flask app to docker
* [test] travis trigger
* [update] readme

This is a simple hello world flask app to develop. The use of alpine is convenient because the images are tiny compared to the size of the official mongo image.

The alpine image is a publicly available, open source image. If the POC is successful I would recommend creating and maintaining our own alpine-mongodb image or at the very least introducing version control rather than pulling the latest version.

This prevents introducing breaking changes - but for the purposes of this exercise it is sufficient.