[add] functionality for `/songs`
This also includes some very basic pagination using the objectId and limiting the number of documents returned at a time. Expected behaviour is to pass in the query string `last-id=xxxxx` and it will return the next 5 documents.

Ideally this would be implemnented with a cursor and include proper page info in the request:

```
{
  "data": {
    "hero": {
      "name": "R2-D2",
      "friendsConnection": {
        "totalCount": 3,
        "edges": [
          {
            "node": {
              "name": "Han Solo"
            },
            "cursor": "Y3Vyc29yMg=="
          },
          {
            "node": {
              "name": "Leia Organa"
            },
            "cursor": "Y3Vyc29yMw=="
          }
        ],
        "pageInfo": {
          "endCursor": "Y3Vyc29yMw==",
          "hasNextPage": false
        }
      }
    }
  }
}
```